Switch to ci user when given custom command
The ENTRYPOINT option is now what is run as root and the configured CMD
(which is overridden by args on the command line) is run as the ci user
after processes have been started. This means that users can run:
  docker run ... freerunning/frt-ci /bin/bash
and be dropped into a shell which is ready for commands to be run.

This switches to using gosu, which ensures that our new process receives
unix signals sent to docker